with_kafka/requirements.txt
amqp==5.1.1
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
kafka-python==2.0.2
kombu==5.2.4
prompt-toolkit==3.0.38
pytz==2022.7.1
six==1.16.0
vine==5.0.0
wcwidth==0.2.6
